requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
Authlib==1.7.0
beautifulsoup4==4.14.3
behave==1.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpickle==3.1.2
codespell==2.4.2
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.7
cssselect==1.4.0
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
docstring_parser==0.18.0
exceptiongroup==1.3.1
execnet==2.1.2
fastapi==0.136.0
fasteners==0.20
fastuuid==0.14.0
feedparser==6.0.12
filelock==3.29.0
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2025.12.0
gcsfs==2025.12.0
google-adk==1.31.1
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-cloud-aiplatform==1.148.1
google-cloud-appengine-logging==1.9.0
google-cloud-audit-log==0.5.0
google-cloud-bigquery==3.41.0
google-cloud-bigquery-storage==2.37.0
google-cloud-bigtable==2.36.0
google-cloud-core==2.5.1
google-cloud-dataplex==2.18.0
google-cloud-discoveryengine==0.13.12
google-cloud-iam==2.22.0
google-cloud-logging==3.15.0
google-cloud-monitoring==2.30.0
google-cloud-pubsub==2.37.0
google-cloud-resource-manager==1.17.0
google-cloud-secret-manager==2.27.0
google-cloud-spanner==3.65.0
google-cloud-speech==2.38.0
google-cloud-storage==2.19.0
google-cloud-storage-control==1.11.0
google-cloud-trace==1.19.0
google-crc32c==1.8.0
google-genai==1.73.1
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
graphviz==0.21
greenlet==3.4.0
grpc-google-iam-v1==0.14.4
grpc-interceptor==0.15.4
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
hf-xet==1.4.3
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.11.0
idna==3.12
importlib_metadata==8.7.1
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
joserfc==1.6.4
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
lxml==6.1.0
Mako==1.3.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.9
mcp==1.27.0
mdurl==0.1.2
mmh3==5.2.1
mplsoccer==1.6.1
multidict==6.7.1
mycdp==1.3.7
nest-asyncio==1.6.0
numpy==2.4.4
oauthlib==3.3.1
openai==2.32.0
opentelemetry-api==1.38.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-google-genai==0.7b0
opentelemetry-proto==1.38.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-genai==0.3b0
orjson==3.11.8
outcome==1.3.0.post0
packaging==26.1
pandas==3.0.2
parameterized==0.9.0
parse==1.21.1
parse_type==0.6.6
pdbp==1.8.2
pillow==12.2.0
platformdirs==4.9.6
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
pyarrow==24.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
PyJWT==2.12.1
pynose==1.5.5
pyOpenSSL==26.0.0
pyotp==2.9.0
pyparsing==3.3.2
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-html==4.0.2
pytest-metadata==3.1.1
pytest-ordering==0.6
pytest-rerunfailures==16.1
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
rich==15.0.0
rpds-py==0.30.0
ruamel.yaml==0.19.1
ruff==0.15.11
sbvirtualdisplay==1.4.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==82.0.1
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
sqlalchemy-spanner==1.17.3
sqlparse==0.5.5
sse-starlette==3.3.4
starlette==0.52.1
tabcompleter==1.4.0
tenacity==9.1.4
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
trio==0.33.0
trio-websocket==0.12.2
ty==0.0.32
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.34.3
watchdog==6.0.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wheel==0.47.0
wrapper-tls-requests==1.2.5
wrapt==1.17.3
wsproto==1.3.2
yarl==1.23.0
zipp==3.23.1
